Collapse project into single script
