--- pbn_manager/__init__.py
@@ -1,11 +1,0 @@
-"""PBN Manager package.
-
-Provides classes for interacting with multiple WordPress sites and
-aggregating statistics. The package is intentionally lightweight and
-relies on the WordPress REST API.
-"""
-
-__all__ = ["WordPressClient", "PBNManager", "WordPressError"]
-
-from .wordpress_client import WordPressClient, WordPressError
-from .manager import PBNManager
